+++ ios/MullvadREST/Relay/RelayPicking/MultihopPicker.swift
@@ -6,10 +6,12 @@
 //  Copyright © 2026 Mullvad VPN AB. All rights reserved.
 //
 
+import MullvadLogging
 import MullvadSettings
 import MullvadTypes
 
 struct MultihopPicker: RelayPicking {
+    let logger = Logger(label: "MultihopPicker")
     let obfuscation: RelayObfuscation
     let tunnelSettings: LatestTunnelSettings
     let connectionAttemptCount: UInt
@@ -26,7 +28,7 @@ struct MultihopPicker: RelayPicking {
             daitaEnabled: daitaSettings.daitaState.isEnabled
         )
 
-        let supportedObfuscation = RelayObfuscator(
+        let supportedObfuscation = try RelayObfuscator(
             relays: obfuscation.allRelays,
             tunnelSettings: tunnelSettings,
             connectionAttemptCount: connectionAttemptCount,
@@ -89,7 +91,7 @@ struct MultihopPicker: RelayPicking {
         relay: SelectedRelay,
         from candidates: [RelayWithLocation<REST.ServerRelay>],
         closeTo location: Location? = nil,
-        applyObfuscatedIps: Bool,
+        applyObfuscation: Bool,
     ) throws -> SelectedRelay {
         let filteredCandidates = candidates.filter { relayWithLocation in
             relayWithLocation.relay.hostname != relay.hostname
@@ -98,7 +100,7 @@ struct MultihopPicker: RelayPicking {
         return try findBestMatch(
             from: filteredCandidates,
             closeTo: location,
-            applyObfuscatedIps: applyObfuscatedIps,
+            applyObfuscation: applyObfuscation,
         )
     }
 }