+++ ios/MullvadVPNUITests/RelayTests.swift
@@ -241,6 +241,64 @@ class RelayTests: LoggedInWithTimeUITestCase {
         try generateTrafficAndDisconnect(from: connectedToIPAddress, searchForPort: 51900, assertProtocol: .UDP)
     }
 
+    func testWireGuardOverLwoCustomPort() throws {
+        try XCTSkipIf(true, "TODO: Enable again when LWO is working as intended everywhere")
+
+        addTeardownBlock {
+            HeaderBar(self.app)
+                .tapSettingsButton()
+
+            SettingsPage(self.app)
+                .tapVPNSettingsCell()
+
+            VPNSettingsPage(self.app)
+                .tapWireGuardObfuscationExpandButton()
+                .tapWireGuardObfuscationOffCell()
+        }
+
+        HeaderBar(app)
+            .tapSettingsButton()
+
+        SettingsPage(app)
+            .tapVPNSettingsCell()
+
+        VPNSettingsPage(app)
+            .tapWireGuardObfuscationExpandButton()
+            .tapWireGuardObfuscationLwoCell()
+            .tapLwoPortSelectorButton()
+
+        LwoObfuscationSettingsPage(app)
+            .tapCustomCell()
+            .typeTextIntoCustomField("4000")
+            .tapBackButton()
+
+        VPNSettingsPage(app)
+            .tapBackButton()
+
+        SettingsPage(app)
+            .tapDoneButton()
+
+        // The packet capture has to start before the tunnel is up,
+        // otherwise the device cannot reach the in-house router anymore
+        startPacketCapture()
+
+        TunnelControlPage(app)
+            .tapConnectButton()
+
+        allowAddVPNConfigurationsIfAsked()
+
+        TunnelControlPage(app)
+            .waitForConnectedLabel()
+
+        let (connectedToIPAddress, _) = TunnelControlPage(app)
+            .tapRelayStatusExpandCollapseButton()
+            .getInIPAddressAndPortFromConnectionStatus()
+
+        try Networking.verifyCanAccessInternet()
+
+        try generateTrafficAndDisconnect(from: connectedToIPAddress, searchForPort: 4000, assertProtocol: .UDP)
+    }
+
     func testWireGuardOverTCPManually() throws {
         addTeardownBlock {
             HeaderBar(self.app)
@@ -346,7 +404,7 @@ class RelayTests: LoggedInWithTimeUITestCase {
 
         VPNSettingsPage(app)
             .tapWireGuardObfuscationExpandButton()
-            .tapWireGuardObufscationQuicCell()
+            .tapWireGuardObfuscationQuicCell()
             .tapBackButton()
 
         SettingsPage(app)
@@ -395,20 +453,20 @@ class RelayTests: LoggedInWithTimeUITestCase {
         try generateTrafficAndDisconnect(from: connectedToIPAddress, searchForPort: 443, assertProtocol: .UDP)
     }
 
-    /// Test automatic switching to TCP is functioning when UDP traffic to relays is blocked.
-    func testWireGuardOverTCPAutomatically() throws {
-        FirewallClient().removeRules()
-        removeFirewallRulesInTearDown = true
+    func testWireGuardOverLwoManually() throws {
+        try XCTSkipIf(true, "TODO: Enable again when LWO is working as intended everywhere")
 
         addTeardownBlock {
-            self.restoreDefaultCountry()
-        }
+            HeaderBar(self.app)
+                .tapSettingsButton()
 
-        // Run actual test
-        try FirewallClient().createRule(
-            // Block all UDP traffic not going to the router.
-            FirewallRule.makeBlockUDPTrafficRule(toIPAddress: "8.8.8.8", inverted: true)
-        )
+            SettingsPage(self.app)
+                .tapVPNSettingsCell()
+
+            VPNSettingsPage(self.app)
+                .tapWireGuardObfuscationExpandButton()
+                .tapWireGuardObfuscationOffCell()
+        }
 
         HeaderBar(app)
             .tapSettingsButton()
@@ -418,25 +476,23 @@ class RelayTests: LoggedInWithTimeUITestCase {
 
         VPNSettingsPage(app)
             .tapWireGuardObfuscationExpandButton()
-            .tapWireGuardObfuscationAutomaticCell()
+            .tapWireGuardObfuscationLwoCell()
             .tapBackButton()
 
         SettingsPage(app)
             .tapDoneButton()
 
         TunnelControlPage(app)
-            .tapSelectLocationButton()
-
-        SelectLocationPage(app)
-            .tapLocationCell(withName: BaseUITestCase.testsDefaultQuicCountryName)
+            .tapConnectButton()
 
         allowAddVPNConfigurationsIfAsked()
 
-        // Should be two UDP connection attempts but sometimes only one is shown in the UI
         TunnelControlPage(app)
-            .tapRelayStatusExpandCollapseButton()
-            .verifyConnectingOverTCPAfterUDPAttempts()
             .waitForConnectedLabel()
+
+        try Networking.verifyCanAccessInternet()
+
+        TunnelControlPage(app)
             .tapDisconnectButton()
     }