+++ ios/PacketTunnelCoreTests/Mocks/ProtocolObfuscationStub.swift
@@ -6,6 +6,8 @@
 //  Copyright © 2026 Mullvad VPN AB. All rights reserved.
 //
 
+import WireGuardKitTypes
+
 @testable import MullvadREST
 @testable import MullvadSettings
 @testable import MullvadTypes
@@ -16,6 +18,7 @@ struct ProtocolObfuscationStub: ProtocolObfuscation {
 
     func obfuscate(
         _ endpoint: MullvadTypes.SelectedEndpoint,
+        clientPublicKey: PublicKey
     ) -> ProtocolObfuscationResult {
         .init(endpoint: endpoint)
     }